refactor: make the api client store look closer to what we have in the reference
