+++ packages/api-client/src/components/ApiClient/RequestHistoryItem.vue
@@ -3,7 +3,7 @@ import prettyBytes from 'pretty-bytes'
 import prettyMilliseconds from 'pretty-ms'
 
 import { humanDiff } from '../../helpers'
-import { useRequestStore } from '../../stores/requestStore'
+import { useRequestStore } from '../../stores'
 import { type ClientResponse } from '../../types'
 
 defineProps<{ history: string }>()
@@ -63,3 +63,4 @@ const getContentLength = (response: ClientResponse) => {
   cursor: default;
 }
 </style>
+../../stores/useRequestStore
