+++ packages/api-client/src/stores/index.ts
@@ -1,2 +1,2 @@
-export { useApiClientStore } from './apiClientStore'
-export { useRequestStore, createEmptyAuthState } from './requestStore'
+export { useApiClientStore } from './useApiClientStore'
+export { useRequestStore, createEmptyAuthState } from './useRequestStore'
