+++ packages/api-reference/src/components/ApiClientModal.vue
@@ -21,9 +21,6 @@ const { hideApiClient, state } = useApiClientStore()
 
 const isMobile = useMediaQuery('(max-width: 1000px)')
 </script>
-<script lang="ts">
-export { useApiClientStore } from '@scalar/api-client'
-</script>
 <template>
   <div
     v-if="state.showApiClient || overloadShow"
